+++ R/rd.R
@@ -7,8 +7,8 @@
 #'
 #' It is run by default by [roxygenize()].
 #'
-#' @seealso [tags-rd], [tags-rd-other], [tags-reuse], [tags-index-crossref] for
-#'   tags provided by this roclet.
+#' @seealso [tags-rd], [tags-rd-data], [tags-rd-oop], [tags-reuse],
+#'   [tags-index-crossref] for tags provided by this roclet.
 #' @export
 #' @examples
 #' #' Add together two numbers
@@ -28,6 +28,8 @@ rd_roclet <- function() {
 
 #' @export
 roclet_process.roclet_rd <- function(x, blocks, env, base_path) {
+  blocks <- merge_external_r6methods(blocks)
+
   # Convert each block into a topic, indexed by filename
   topics <- RoxyTopics$new()